実験用３つ受信

+++ include/udp_connect.hpp
@@ -30,11 +30,11 @@ class UdpConnect{
 
         UdpConnect(std::string address, int port, size_t element_count); // UDPコンストラクタ       
         
-        void udp_send(const std::vector<double>& values , int roop_count); // UDP送信関数
+        void udp_send(const std::vector<double>& values , long roop_count); // UDP送信関数
 
         void udp_bind(); 
 
-        std::pair<std::vector<double>, int> udp_recv();
+        std::pair<std::vector<double>, long> udp_recv();
 
         
         ~UdpConnect();
